Added snapshot-id translation for alternative storage implementations
Snapshot identifications may differ for each storage implementation.
To overcome this problem we create a SnapshotIndex that stores in

the operation Snapshot the returned ID.
When replaying we associate the recorded ID with the replayed ID so
that the revertSnapshot operation can revert to the snapshot with
the replayed ID (not the recorded one).