Add Code Dictionary
This PR adds a code dictionary to Aida so that the bytecode of a SetCode instruction
can be stored in the trace. The dictionary is a simple map that converts a bytecode
to a unique ordinal number. Note that the internal representation of bytecode is a
string not a slice of byte (due to the map inside the dictionary).