Add GetCode/GetCodeSize Operations
This PR adds the operations GetCode() and GetCodeSize() to the recorder
and replayer. At the moment, the StateDB has not implemented both operations.
As soon as they are implemented in the interface, we need to uncomment
the code blocks in the Execute() methods so that both operations are really
executed in the replayer.