Adds AddBalance & SubBalance operations
This PR adds AddBalance and SubBalance operation for the trace
recorder/replayer. The operations keep amount in a fixed length
16-byte array (amount doesn't fit in 8 bytes).

Execution() of SubBalance is commented out because executing
SubBalance on an empty statedb returns a negative balance which
causes a runtime error in Finalise(). Once statedb priming is
completed, Execution() of SubBalance can be uncommented.