Optimize Trace Iterator
The trace iterator had un-buffered I/O which was too slow to read
large quantities. The file seek operations for checking the
last block in the block range was too slow as well.

This Pull-Request enables buffered I/O and uses BeginBlock to check
whether the last block of the block range has been reached.