Compress Trace File
The current trace file is written and read in an uncompressed format that increases in size too quickly with the number of blocks recorded.

Using bzip2 at level 9 (the best level) gives a compression factor of approximately 5.3 (to an uncompressed file format) for the blocks from 5M to 5.2M on mainnet.

For 200K blocks, we have a trace file size of 115.65 MiB. This means we need approx. 6KiB per block.