Reset storage of deleted accounts
This PR is a part of a fix which clears storage of suicided contracts.
Since substate only keeps touched addresses, it cannot distinguish
between storage keys have been deleted or they are not accessed. For
simplicity, a storage slot is marked as deleted by setting to
common.Hash{}. This change is particularly important for accounts which
later resurrected.

This PR includes
- reset storage of suicided accounts when advancing state
- improve interface of worldstate update