Stochastic Replay
This PR provides an initial version of the stochastic replayer that does not
connect to a StateDB database. Instead it prints StateDB operation on the screen
if the versboe flag is set.

Several features such as the connection to the StateDB and simulation of some
parameters (such as balance, nonce, snapshots) is missing.