Fix stationary distribution computation
This PR fixes the stationary distribution (aka. steady-state). There was
the underlying assumption that the eigenvalue of one is always the first index.
This assumption was wrong and the eigenvalue of one must be searched in the
vector of eigenvalues. When found, the associated left-eigenvector needs to be
normalized and returned as stationary distribution.