Fix the compression of the counting statistics
This PR changes the algorithm for compressing the empirical cumulative distribution function using the
Visvalingam-Whyatt algorithm. The algorithm permits to use many less points in the emperical
cumulative distribution function by removing the least important points. The least important point
is a point whose triangle area formed by the point and its adjacent points is the smallest.

More details on the algorithm can be found here:
https://en.wikipedia.org/wiki/Visvalingam%E2%80%93Whyatt_algorithm