Fix StateDB connection for stochastic replay
This PR establishes a StateDB connection to the stochastic simulation.
The main challenge for the connection was the simulation of the snapshot
mechanism so that the actual balance is known at a snapshot.
Without creating a balance log, the simulation may selects a value for
the SubBalance operation, so that the account balance becomes negative
resulting in a StateDB crash.