Record deleted accounts and resurrected accounts
This PR is a part of a fix which clears storage of suicided contracts.
Since substate only keeps touched addresses, it cannot distinguish
between storage keys have been deleted or they are not accessed.
For simplicity, a storage slot is marked as deleted by setting to
common.Hash{}. This change is particularly important for accounts
which later resurrected.

Recording of deleted addresses and resurrected addresses

Deleted address: an account which was successfully suicided in a transaction.
Resurrected address: an account created in a transaction which was previously
suicided. Only successful transactions are recorded.