Refactor stochastic cli
This PR separates the estimation from the visualization. The benefit from the separation
is that the after the estimation we can be sure that the simulation file is written and
we don't need to wait for terminating the web-server for visualization (in case the
visualize flag was set).

With this PR we have two new commands
 aida-stochastic estimate <event-file>
 aida-stochastic visualize <event-file>

Both commands use as an argument the event file that is produced by the record command.