Add random-seed for the replay tool
This PR adds a random-seed flag to the stochastic replay tool. This flag permits
to configure the random number generator so that the generated operations are
deterministic (assuming that the DB implementations don't use the random
generator on their own/especially in concurrent threads).

I also removed the Chain-Id flag from the replay tool which required to set the
internal chaind-id to the default setting (otherwise it receives the zero value
and the replay tool terminates).