Remove output channel and write directly
This PR removes the output channel in the record part of the tracer to
avoid termination races for the close() operation of files. The StateDB
operations are written directly to file. If the buffer sizes are set
to a considerable size, it will not hurt performance.

In addition, it simplifies the code with less threads (avoiding data
races).