Fix deletion index
The DeleteIndex() methods already receives a shifted value
and does not decrement the value by one.
In addition, the deletion only neeeds to be performed if
the last element in the index range is deleted.
The error result for the DeleteIndex method was not checked
in the replay.

This PR addresses aforementioned issues.