Isolate use of random generator
This PR isolates the use of the random generator for the stochastic tool. The stochastic
tool may produce indeterministic sequences of StateDB operations if there are other
consumers for system random generator.

A new random generator is created and passed through the components of the system that
require a random generator. With that change, the sequence becomes deterministic since
there can be no interference from other components in the system (using the random
generator).