Fix cli arguments for generate in aida-stochastic
Currently, aida-stochastic generate requires a first and last block. However, for the command generate this makes sense (there is no first and last block required for generating an event file).

This PR adds a NoArgs mode for the configuration environments to accommodate for tools that do not require substate/stateDB to be opened.