Non-Interfering implementation of GetBalance() for stochastic replay (#485)
* Non-Interfering implementation of GetBalance() for stochastic replay

This PR extends the statedb interface so that the shadow db implementation becomes accessible. This is necessary
for avoiding potential interference, i.e., before SubBalance() issued by stochastic a GetBalance() on the same
DB (System under Test) is issued. This may cause an interference. By having accessing to the shadow DB (in case
there is one specified), the GetBalance() operation can be executed on the shadow DB.

If no shadow DB is specified, the GetBalance() operation is executed on the primary DB.

* Fix test cases

* Change GetShadowDB() semantics to return nil if no shadow DB exists