Parallel Experiment Report
This PR adds a script to run the parallel experiment (cf #547), adds a R rendering scripts for R Markdown documents,
and adds a report in R Markdown for presenting the data obtained by the aida-profile parallelisation tool.

The parameters of the parallel experiment script are as follows:

run_parallel_experiment.sh <aida-db> <carmen-impl> <carmen-variant> <tosca-impl> <tmp-directory> <startBlock> <endBlock> <output-directory>

The script must run in the main directory of the Aida repository.