Remove world state cli
This change remove uses of world-state cli and its library.
Update set reads the initial world state from WorldStateDB which is
compressed in its own format. The functions is currently removed in
this PR. However, the data type will be change such that the WorldStateDB
should contain data in updateset structure instead.